requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
cachetools>=5.3.0
aiofiles>=23.2.0
reportlab>=4.0.0
pyswisseph>=2.10.3
geopy>=2.4.0
pytz>=2024.1
